+++ internal/src/core/standard/service/impl/standard.py
@@ -8,8 +8,9 @@ from internal.src.core.show.service.show import IShowService
 from internal.src.core.standard.repository.standard import IStandardRepository
 from internal.src.core.standard.schema.standard import StandardSchema, StandardSchemaCreate, StandardSchemaDeleteResponse
 from internal.src.core.standard.service.standard import IStandardService
-from internal.src.core.utils.exceptions import StandardServiceError
-from internal.src.core.utils.types import ID, Weight
+from internal.src.core.utils.exceptions import StandardServiceError, CheckAnimalStandardError, CheckAnimalBreedError, \
+    NotFoundRepoError, StandardInUseError
+from internal.src.core.utils.types import ID, Weight, Height, Length
 
 
 class StandardService(IStandardService):
@@ -33,40 +34,55 @@ class StandardService(IStandardService):
         new_standard = StandardSchema.from_create(standard_create)
         return self.standard_repo.create(new_standard)
 
-    def delete(self, id: ID) -> StandardSchemaDeleteResponse:
+    def is_standard_used(self, id: ID) -> bool:
         try:
             self.show_service.get_by_standard_id(id)
-        except HTTPException:
-            self.standard_repo.delete(id)
-            return StandardSchemaDeleteResponse(id=id)
-        raise StandardServiceError(detail=f'standard\'s in use: standard_id={id}')
+        except NotFoundRepoError:
+            return False
+        return True
 
-    def check_animal_by_standard(self, standard_id: ID, animal: AnimalSchema) -> bool:
+    def delete(self, id: ID) -> StandardSchemaDeleteResponse:
+        if self.is_standard_used(id):
+            raise StandardInUseError(standard_id=id)
+
+        self.standard_repo.delete(id)
+        return StandardSchemaDeleteResponse(id=id)
+
+    @staticmethod
+    def check_animal_weight(animal: AnimalSchema, standard: StandardSchema):
+        lo_weight = standard.weight * (100 - standard.weight_delta_percent) / Weight(100)
+        hi_weight = standard.weight * (100 + standard.weight_delta_percent) / Weight(100)
+        if animal.weight < lo_weight or animal.weight > hi_weight:
+            raise CheckAnimalStandardError(animal_id=animal.id, standard_id=standard.id, property_name='weight')
+
+    @staticmethod
+    def check_animal_length(animal: AnimalSchema, standard: StandardSchema):
+        lo_length = standard.length * (100 - standard.length_delta_percent) / Weight(100)
+        hi_length = standard.length * (100 + standard.length_delta_percent) / Weight(100)
+        if animal.length < lo_length or animal.length > hi_length:
+            raise CheckAnimalStandardError(animal_id=animal.id, standard_id=standard.id, property_name='length')
+
+    @staticmethod
+    def check_animal_height(animal: AnimalSchema, standard: StandardSchema):
+        lo_height = standard.height * (100 - standard.height_delta_percent) / Weight(100)
+        hi_height = standard.height * (100 + standard.height_delta_percent) / Weight(100)
+        if animal.height < lo_height or animal.height > hi_height:
+            raise CheckAnimalStandardError(animal_id=animal.id, standard_id=standard.id, property_name='height')
+
+    def check_animal_by_standard(self, standard_id: ID, animal: AnimalSchema):
         cur_standard = self.standard_repo.get_by_id(standard_id)
 
         if cur_standard.breed_id != animal.breed_id:
-            raise StandardServiceError(detail=f'breed_id must be equal: standard_id={standard_id},'
-                                              f' animal_id={animal.id.value}')
+            raise CheckAnimalBreedError(animal_id=animal.id, standard_id=standard_id)
 
-        lo_weight = cur_standard.weight * (100 - cur_standard.weight_delta_percent) / Weight(100)
-        hi_weight = cur_standard.weight * (100 + cur_standard.weight_delta_percent) / Weight(100)
-        if animal.weight < lo_weight or animal.weight > hi_weight:
-            return False
-
-        lo_height = cur_standard.height * (100 - cur_standard.height_delta_percent) / Weight(100)
-        hi_height = cur_standard.height * (100 + cur_standard.height_delta_percent) / Weight(100)
-        if animal.height < lo_height or animal.height > hi_height:
-            return False
-
-        lo_length = cur_standard.length * (100 - cur_standard.length_delta_percent) / Weight(100)
-        hi_length = cur_standard.length * (100 + cur_standard.length_delta_percent) / Weight(100)
-        if animal.length < lo_length or animal.length > hi_length:
-            return False
+        self.check_animal_weight(animal, cur_standard)
+        self.check_animal_height(animal, cur_standard)
+        self.check_animal_length(animal, cur_standard)
 
         if animal.has_defects != cur_standard.has_defects:
-            return False
+            raise CheckAnimalStandardError(animal_id=animal.id, standard_id=standard_id, property_name='has_defects')
 
         if animal.is_multicolor != cur_standard.is_multicolor:
-            return False
+            raise CheckAnimalStandardError(animal_id=animal.id, standard_id=standard_id, property_name='is_multicolor')
 
         return True
